Added support for master server reinstalls.

+++ rimuapi.py
@@ -1,6 +1,7 @@
 import urllib
 import os
 from requests import Request, Session
+from warnings import catch_warnings
 
 try:
     import json
@@ -15,6 +16,7 @@ def sort_unique(sequence):
     return itertools.imap(
         operator.itemgetter(0),
         itertools.groupby(sorted(sequence)))
+
 
 
 def valid_domain_name(domain_name):
@@ -51,7 +53,6 @@ class Api:
 
         self._distros = []
         self._plans = []
-        self._orders = []
 
         if not self._key:
             self._key = os.getenv('RIMUHOSTING_APIKEY', None)
@@ -60,16 +61,17 @@ class Api:
             if settings:
                 self._key = settings.apikey
 
-    def __send_request(self, url, data=None, method='GET'):
-        if self._key:
-            headers = {
-                'Authorization': "rimuhosting apikey=%s" % self._key,
-                'Content-Type': 'application/json',
-                'Accept': 'application/json'
-            }
-        else:
-            raise Exception('Get API Key')
-        url = urllib.basejoin(self._base_url, url)
+    def __send_request(self, url, data=None, method='GET', isKeyRequired=True):
+        if isKeyRequired and not self._key:
+            raise Exception('API Key is required.  e.g. export RIMUHOSTING_APIKEY=xxxx .  Get the API key from http://rimuhosting.com/cp/apikeys.jsp')
+        headers = {
+            'Content-Type': 'application/json',
+            'Accept': 'application/json'
+        }
+        if isKeyRequired:
+            headers['Authorization']= "rimuhosting apikey=%s" % self._key
+        
+        url = urllib.parse.urljoin(self._base_url, url)
 
         data = data if isinstance(data, str) else json.dumps(data)
 
@@ -80,57 +82,80 @@ class Api:
                       )
         prepped = s.prepare_request(req)
         resp = s.send(prepped)
+        if not resp.ok:
+            message = resp.text
+            try: 
+                message = resp.json()["jaxrs_response"]["error_info"]["human_readable_message"]
+            finally:
+                raise Exception(resp.status_code, resp.reason, message)
+        
         return resp
 
     # list available distros
     def distros(self):
-        r = self.__send_request('/r/distributions')
-        if r.ok:
-            data = r.json()
-            self._distros = data['get_distros_response']['distro_infos']
-            return self._distros
-        else:
-            raise Exception(r.status_code, r.reason)
+        r = self.__send_request('/r/distributions', isKeyRequired=False)
+        data = r.json()
+        self._distros = data['get_distros_response']['distro_infos']
+        return self._distros
 
     # list pricing plans & data centers
     def plans(self):
-        r = self.__send_request('/r/pricing-plans')
-        if r.ok:
-            data = r.json()
-            self._plans = data['get_pricing_plans_response']['pricing_plan_infos']
-            return self._plans
-        else:
-            raise Exception(r.status_code, r.reason)
+        r = self.__send_request('/r/pricing-plans', isKeyRequired=False)
+        data = r.json()
+        self._plans = data['get_pricing_plans_response']['pricing_plan_infos']
+        return self._plans
 
     def data_centers(self):
+        import itertools
         try:
             plans = self._plans
         except AttributeError:
             plans = self.plans()
-        data_centers = list(sort_unique([i['offered_at_data_center'] for i in plans]))
-        return data_centers
+        dcs = []
+        lookup = {}
+        from pprint import pprint
+
+        for i in plans:
+            i = i['offered_at_data_center']
+            if not i: 
+                continue
+            code = i['data_center_location_code']
+            if not code:
+                continue
+            if not code in lookup:
+                lookup[code] = i;
+                dcs.append(i);
+        return dcs; 
 
     # list of orders/servers
-    def orders(self, include_inactive='N'):
-        payload = {'include_inactive': include_inactive}
-        r = self.__send_request('/r/orders;%s' % urllib.urlencode(payload))
-        if r.ok:
-            data = r.json()
-            self._orders = data['get_orders_response']['about_orders']
-            return self._orders
-        else:
-            raise Exception(r.status_code, r.reason)
+    def orders(self, include_inactive='N', filter={}):
+        filter['include_inactive'] = include_inactive
+        uri = '/r/orders;%s' % urllib.parse.urlencode(filter)
+        uri = uri.replace('&', ';')
+        r = self.__send_request(uri)
+        data = r.json()
+        print("order search uri of " + str(uri) + " returns " + str(data))
+        return data['get_orders_response']['about_orders']
 
-    # create server
-    def create(self, domain, **kwargs):
+    def _get_req(self, domain=None, kwargs={}):
         _options, _params, _req = {}, {}, {}
-        if not valid_domain_name(domain):
+        _req = kwargs
+        if not 'instantiation_options' in _req:
+            _req['instantiation_options'] = _options
+        if not 'vps_parameters' in _req:
+            _req['vps_parameters'] = _params
+        _options = _req['instantiation_options']
+        _params = _req['vps_parameters']
+        if domain:
+            _options['domain_name'] = domain
+        if not valid_domain_name(_options['domain_name']):
             raise Exception(418, 'Domain not valid')
-        _options['domain_name'] = domain
         if 'password' in kwargs:
             _options['password'] = kwargs['password']
         if 'distro' in kwargs:
             _options['distro'] = kwargs['distro']
+        if 'cloud_config_data' in kwargs:
+            _options['cloud_config_data'] = kwargs['cloud_config_data']
         if 'control_panel' in kwargs:
             _options['control_panel'] = kwargs['control_panel']
         if 'disk_space_mb' in kwargs:
@@ -144,79 +169,57 @@ class Api:
         if 'file_injection_data' in kwargs:
             _req['file_injection_data'] = kwargs['file_injection_data']
         if 'ssh_pub_key' in kwargs:
-            _req['file_injection_data'] = []
+            if not _req['file_injection_data']:
+                _req['file_injection_data'] = []
             _req['file_injection_data'].append(
                 {'data_as_string': kwargs['ssh_pub_key'],
                  'path': '/root/.ssh/authorized_keys'})
-        _req['instantiation_options'] = _options
-        _req['vps_parameters'] = _params
+        return _req
+    
+    # create server
+    def create(self, domain, **kwargs):
+        _req = self._get_req(domain, kwargs)
         payload = {'new_order_request': _req}
         r = self.__send_request('/r/orders/new-vps', data=payload, method='POST')
-        if r.ok:
-            return r.json()
-        else:
-            raise Exception(r.status_code, r.reason, r.text)
+        return r.json()
+
+    def create(self, vmargs={}):
+        _req = self._get_req(domain=None, kwargs=vmargs)
+        payload = {'new_order_request': _req}
+        r = self.__send_request('/r/orders/new-vps', data=payload, method='POST')
+        return r.json()
 
     # reinstall server
     def reinstall(self, domain, order_oid, **kwargs):
-        _options, _params, _req = {}, {}, {}
-        if not valid_domain_name(domain):
-            raise Exception(418, 'Domain not valid')
-
-        _options['domain_name'] = domain
-        if 'password' in kwargs:
-            _options['password'] = kwargs['password']
-        if 'distro' in kwargs:
-            _options['distro'] = kwargs['distro']
-        if 'control_panel' in kwargs:
-            _options['control_panel'] = kwargs['control_panel']
-        if 'disk_space_mb' in kwargs:
-            _params['disk_space_mb'] = kwargs['disk_space_mb']
-        if 'memory_mb' in kwargs:
-            _params['memory_mb'] = kwargs['memory_mb']
-        if 'dc_location' in kwargs:
-            _req['dc_location'] = kwargs['dc_location']
-        if 'meta_data' in kwargs:
-            _req['meta_data'] = kwargs['meta_data']
-        if 'file_injection_data' in kwargs:
-            _req['file_injection_data'] = kwargs['file_injection_data']
-        if 'ssh_pub_key' in kwargs:
-            _req['file_injection_data'] = []
-            _req['file_injection_data'].append(
-                {'data_as_string': kwargs['ssh_pub_key'],
-                 'path': '/root/.ssh/authorized_keys'})
-        _req['instantiation_options'] = _options
-        _req['vps_parameters'] = _params
+        _req = self._get_req(domain, kwargs)
         payload = {'new_order_request': _req}
         r = self.__send_request('/r/orders/order-%s-%s/vps/reinstall' % (order_oid, domain),
                                 data=payload, method='PUT')
-        if r.ok:
-            return r.json()
-        else:
-            raise Exception(r.status_code, r.reason, r.text)
+        return r.json()
+
+    def reinstall(self, order_oid, vmargs={}):
+        _req = self._get_req(domain=None, kwargs=vmargs)
+        payload = {'new_order_request': _req}
+        r = self.__send_request('/r/orders/order-%s-%s/vps/reinstall' % (order_oid, "na.com"),
+                                data=payload, method='PUT')
+        return r.json()
 
     # check status
     def status(self, domain, order_oid):
         r = self.__send_request('/r/orders/order-%s-%s/vps' % (order_oid, domain))
-        if r.ok:
-            data = r.json()
-            return data['get_vps_status_response']['running_vps_info']
-        else:
-            raise Exception(r.status_code, r.reason, r.text)
+        data = r.json()
+        return data['get_vps_status_response']['running_vps_info']
 
     def info(self, domain, order_oid):
         r = self.__send_request('/r/orders/order-%s-%s' % (order_oid, domain))
-        if r.ok:
-            data = r.json()
-            return data['get_order_response']['about_order']
-        else:
-            raise Exception(r.status_code, r.reason, r.text)
+        data = r.json()
+        return data['get_order_response']['about_order']
 
-    def _get_order_oid(self, domain=None, ip=None):
+    def _get_order_oid(self, domain=None, ip=None, orders=None):
         oids = []
         if domain is None and ip is None:
             return False
-        orders = self._orders if self._orders else self.orders()
+        orders = orders if orders else self.orders()
         for o in orders:
             if o['domain_name'] == domain:
                 oids.append(o['order_oid'])
@@ -225,16 +228,15 @@ class Api:
         return oids
 
     # cancel server
-    def delete(self, domain, order_oid):
+    def delete(self, domain="na.com", order_oid=0):
+        if valid_domain_name(domain) and order_oid <1:
+            raise Exception("Need an order id")
         if not valid_domain_name(domain):
             raise Exception(418, 'Domain not valid')
         else:
             r = self.__send_request('/r/orders/order-%s-%s/vps' % (order_oid, domain),
                                     method='DELETE')
-            if r.ok:
-                return r.json()
-            else:
-                raise Exception(r.status_code, r.reason, r.text)
+            return r.json()
 
     # change state of server
     # states: RUNNING | NOTRUNNING | RESTARTING | POWERCYCLING
@@ -244,10 +246,7 @@ class Api:
         r = self.__send_request('/r/orders/order-%s-%s/vps/running-state' % (order_oid, domain),
                                 data={'running_state_change_request': {'running_state': new_state}},
                                 method='PUT')
-        if r.ok:
-            return r.json()
-        else:
-            raise Exception(r.status_code, r.reason)
+        return r.json()
 
     def reboot(self, domain, order_oid):
         return self.change_state(domain, order_oid, 'RESTARTING')
@@ -274,7 +273,4 @@ class Api:
                                     'pricing_change_option': pricing_change_option,
                                     'selected_host_server_oid': selected_host_server_oid}},
                                 method='PUT')
-        if r.ok:
-            return r.json()
-        else:
-            raise Exception(r.status_code, r.reason)
+        return r.json()
